<==== Mercurial data ====>
Branch: cloud_19.1

+++ vms/libs/common/src/analytics/detected_objects_storage/analytics_events_receptor.cpp
@@ -38,7 +38,8 @@ void AnalyticsEventsReceptor::putData(const QnAbstractDataPacketPtr& data)
     bool isParsedSuccessfully = false;
     *detectionMetadataPacket =
         QnUbjson::deserialized<nx::common::metadata::DetectionMetadataPacket>(
-            QByteArray::fromRawData(metadataPacket->data(), metadataPacket->dataSize()),
+            QByteArray::fromRawData(metadataPacket->data(),
+            (int) metadataPacket->dataSize()),
             nx::common::metadata::DetectionMetadataPacket(),
             &isParsedSuccessfully);
     if (!isParsedSuccessfully)
